Moved interfaces to new package to clean up explorer view.
